Complete P9.11 DHMS

+++ P9.10 Minutes and seconds/Program.cs
@@ -5,7 +5,7 @@ Give me a number that represents seconds:");
 string answer1 = Console.ReadLine();
 int s = int.Parse(answer1);
 
-float m = s / 60;
+int m = s / 60;
 
 int r = s % 60;
